test: Move `which` method to helper
We use which in other tests, so we moved it to the helper.
