+++ test/helpers/sandbox.rb
@@ -134,6 +134,17 @@ module Helpers
     end
   end
 
+  module LinuxCommand
+    def which(command)
+      ENV["PATH"].split(File::PATH_SEPARATOR).each do |path|
+        next unless File.absolute_path?(path)
+        absolute_path_command = File.join(path, command)
+        return absolute_path_command if File.executable?(absolute_path_command)
+      end
+      nil
+    end
+  end
+
   module PlatformDetectable
     private
     def windows?
